third_party/blink/web_tests/http/tests/misc/onload-detach-during-csp-frame-src-none-expected.txt

CONSOLE ERROR: Framing '' violates the following Content Security Policy directive: "frame-src 'none'". The request has been blocked.

PASS
